[common] KUB-6: Try and prevent orphaned buckets and IAM users
If a provision fails, we can end up with orphaned buckets, because
the next attemt will use a different random uuid for the bucket suffix.

Ideally we'd be able to return the state prior to the failure to the object
library, together with an indication on whether it's worth retrying the
provisioning. Then we could continue where we left off on the next attempt to provision. But that would be a lot of work on a deprecated library.

Instead, use defers in the provisioning to undo completed actions if the provisioning fails. This will not help if the s3/iam endpoints go offline during the operation, preventing the provisioning completing and being able to undo bucket/iam user creation. So it is not a complete solution.